#79 [RENAME] 사이드바 커피챗/그룹챗 리스트 파일 이름 변경(일반 리스트 조회와 구분)
