+++ src/pages/GroupChatPage/hooks/useFetchCoffeeChat.ts
@@ -1,4 +1,4 @@
-import { fetchCoffeeChat } from "@/pages/GroupChatPage/apis/fetchCoffeeChat";
+import { fetchCoffeeChat } from "@/pages/GroupChatPage/apis/fetchSideCoffeeChatList";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { useParams } from "react-router-dom";
